Adding api.js and defining methods there for registering and showing all users. Also moving DB stuff to separate files to clean up app.js.
New functionality:
- when registering a user, you'll get redirected to a page that shows you all users.
