+++ DemoApp/models/user.js
@@ -26,6 +26,4 @@ userSchema.pre('save', function(next) {
       next();
 });
     
-var User = mongoose.model('User', userSchema);
-
-module.exports = User;
+module.exports = mongoose.model('User', userSchema);
