--- DemoApp/routes/layout.jade
@@ -1,17 +1,0 @@
-doctype html
-html
-  head
-    block head 
-        block title
-            title Demo app
-            link(rel='stylesheet', href='/bootstrap/css/bootstrap.min.css')
-            link(rel='stylesheet', href='/bootstrap/css/bootstrap-responsive.min.css')
-            link(rel='stylesheet', href='/stylesheets/style.css')
-            script(src='https://ajax.googleapis.com/ajax/libs/jquery/1/jquery.min.js')
-            script(src='/bootstrap/js/bootstrap.min.js')
-    body
-        include header 
-        
-        block content 
-        
-        include footer
